moved tests to cases folder
